feat(site): move Default Roles UI from Settings to Roles page
Relocate the Default Roles section, its dialog, and the roles query
to the organization Roles page. The selector sits above Custom Roles,
shares the page-level roles query, and stays gated on the experiment
plus the editSettings permission.
